move execution graph to it's own file, rename a few things
